Changes to options
Output options that related to appearance (and also affected browser output) have been moved to Appearance section. Have renamed buttons "Preview" and "Download" to represent the browser preview and the output of selected file. This is partly in anticipation of some client side generation of output at a future date.