Respect XREF setting in Vesta
When using the Vesta suite of modules, you can enable an option to show the XREF in the person's name. This means the option to show XREF no longer works.
In this commit, we make a change so that the GVExport option will successfully disable this, but we also change it so that if the option to show XREF in name is enabled in Vesta (or another similar module), then the option to display the XREF on individuals is turned on by default (otherwise it is off by default).