Add "Anybody" relation option (#86)
* Add "Anybody" relation option

This option will include any linked person in the selected generations (respecting max ancestors/descendants settings). In practice, this adds the relatives of spouses.

* Respect disabling of ancestors/descendants options

* Add parent of unrelated child

Update to fix parent of unrelated child not showing when "Anyone" option set

* Correctly trace tree for "Anyone" option

This update brings a change to how the tree is traversed to support the "Anyone" option, that will traverse anyone know matter how distantly related.

* Fix unrelated child's missing parents

* Change non-blood related colours

Making them slightly more vivid so they can be seen better

* Don't enable "anyone" option by default.

This change disables the "Anyone" option by default.

* Fix partner or partner not included

Small change to include spouses of spouses if "Anyone" option selected

* Don't override parents value

With "Anyone" option there is a lot more rechecking of records. If family record already has "has parents" set to true, don't override it and set it to false just because of the path it was found the second time.
Resolves issue with combined mode sometimes not showing a family and instead just showing the family ID.