Fix grandparents not being shown for adopted starting indi
When starting indi is adopted, grandparents were not shown for mother if "spouses" option checked, but worked fine otherwise. This was because indi's father was added, then the spouse of father, then mother. Because mother already existed, and "anyone" option not set, link to grandparents not followed. Updated to ignore spouse link if adopted setting set to "BOTH", as we know mother will be added by mother link.