Check related on full tree when option set
Currently, when the related option is set, this only checks the displayed records. Where the link is not in the displayed records, the related status is shown as not related. This update will scan the full tree for relatives of starting persons, in order to determine whether an individual is related to the starting person.