Limit when full tree scan runs
This commit limits when the full tree scan for relations happens. Previously, it would run when the option was selected to colour non-relatives differently. Now, it only runs when a non-relative is found in the tree.

This change also fixes some issues with the relative/non-relative check to more accurately show the colouring.