Tidy hover text
When hovering over a family record, a messy version of the contents of the record was shown. When hovering over a URL on an individual, it showed "<TABLE>". It also showed "WT_Graph" when hovering in blank space. This now shows nothing when in blank space, and the XREF of the record when hovering over a record.