Fix dpi (#155)
* Convert var declarations to let/const

Update var to let/const - no change otherwise

* Fix viewBox when DPI over 72

Fix the viewBox when DPI too high
viewBox is set by Viz.js so we edit this after SVG generated, before appending element to the page

* Increase max zoom

We double the max zoom to better handle lower DPI settings

* Remove invalid units

* Fix toast centering for large image

* Add error when canvas too big

If exporting file uses canvas (e.g. JPEG, PNG) then the browser has a max canvas size that varies by browser and in most cases will result in a blank canvas and an invalid zero kilobyte file downloaded. If the output is very short we have hit this limit, so we show an error message.

* Wording change error message

Update error message wording to suggest SVG as an option to avoid canvas if browser canvas limit reached

* Further fix for toast width

Fix for when multiple toasts of different widths are displayed to prevent larger toasts stretching the smaller ones