Remember output type (#161)
After generating the browser output, the output file type was always set to "dot" due to the output being changed to dot by javascript in order to have the dot string returned for displaying in the browser. This changes this so that the output type is not changed, and instead the backend will use "dot" as a value if the $browser variable is true but not change the form or save dot in the cookie.