Make page loaded check more robust
When we run the browser render on page load, occasionally this is still failing if things load in an unusual order.
Additional check added to ensure that the pageLoaded function exists (internal script) and that jQuery exists (external, provided by webtrees) and if not, wait longer before trying to trigger the page loaded script.