Fix scroll to person for combined mode
Combined mode does not hold individual IDs so could not scroll to these records when starting person clicked on.
This change updates the node name for family records in combined mode to have the XREF of the individuals contained. A small change is made to support this change in the function to scroll to a person.