Scale images for server side download PDF & SVG
When inserting images (photos) into PDF and SVG diagrams generated when Graphviz is installed on the server, these are now scaled down first. This significantly reduces file size, while generally not affecting quality as images are displayed reasonably small to start with. Quality is based on DPI so this can be used to further influence the image scaling.