Shorten setting names
webtrees has a limit of approx 32 characters for the name when saving preferences. To implement multiple saved settings, we need to add a suffix, and to do this we need more space.
We are reducing the setting name length to a maximum of 21 characters so the 4 character prefix plus _ to separate allows 6 characters for a saved settings group ID.