Fix issue with rouge records in preferences
Some setting types would be saved to preferences when save button pressed, but not deleted when preference records deleted. This was due to them being part of the form submission code but not part of the default settings. We add these two records to the list of records not saved so solve this issue.