User settings in one filed per tree
User saved settings now are in one preference for each tree. i.e. the preference is saved with a tree and user prefix, e.g. t1_u3, where 1 is the tree ID and 3 is the user ID.

We retain the ID list in a separate field, for more efficient generating of a new ID. One list of settings IDs is kept for each tree, but the IDs are shared across users.