Fix person not showing on stop and indi lists
Duplicate removal test was not adding a person to the list if they showed up anywhere on the page. When list was expanded to have a list of stop persons, this was not updated. So persons were not able to be shown on both lists (though it functioned fine in terms of displaying on the diagram). This resolves the issue.