Resolve issue with case in IDs
As per #329 if you enter a lower case ID the stop person does not work but the person is loaded to the list ok.

XREFs are not case sensitive, so solution is to upper case the list.

Also made change to pulling of person data from webtrees, checking that the returned XREF is the same as searched for. This resolves an issue where another record could be returned if the XREF also appeared in a name or other searchable field.