Don't hide sidebar
When loading settings from a URL token, if the settings have auto-update disabled then don't hide the settings panel.

Also includes some minor refactoring, and a change to loading the token as it was unreliable as to whether the diagram would be loaded when the page loaded if auto-update was off. It would still update if auto-update was on in the saved form (i.e. last time form was loaded). This change fixes this.

Resolves #346