Ignore debug settings if panel not showing
Currently if a user has debug settings saved, these are loaded even if the debug panel is currently disabled. This can cause confusion if the settings are different than expected and can't be changed.

This change updates the functionality to ignore the user setting and load from the admin settings for debug mode and graphviz settings.