Rename new settings to avoid colour/color
Renamed the settings added since the latest release to avoid using either colour or color. Code base is in UK English but we should avoid tying our settings to a particular dialect if possible. Some existing settings have "colour" but these are not changed as an update would affect existing installs.