Fix issue when thumbnails disabled
When the webtrees tree has the show thumbnails setting disabled (which hides images), the diagram was failing to show. This is because having this setting disabled means we don't show the photo options, but the javascript was looking for an element related to a photo option.

This is now updated to handle this case.