Refactor scroll to XREF
Updated scroll to XREF to now take position information. Tile for the XREF will be placed at this position on the screen instead of in the middle.

Updated diagram so if an individual is updated via clicking on them in diagram, then the update will not switch to loading screen and individual clicked on will be in same position on screen and at same zoom level after the update.