Resolve issue with indi list
The lists of individuals, specifically the starting indi and stopping indi lists, have the items show as too large when there is a long name that wraps to the next line. This commit solves this issue by preventing this wrap and instead cutting off the content if it's too long.