Add option to bypass using alt events
Added option "Use similar events if information not available". This is enabled by default and is the current state - webtrees currently uses dates and places of christening/baptism for a date of birth if the DOB is not known. And a date and place of cremation/burial for deaths if not known.

This new option allows for this to be disabled, and only a birth or death event is used.