Adjust page link option
Adjusted option to add links to individuals and families. It previously stated that links were in browser, SVG, and PDF however this is not true.

Browser output has been changed so the behaviour is based on the specific option for what happens when a user clicks a record.

Links in SVG output have been broken for many years, and it seems you cannot have both images and links because different Graphviz renderers are used for each.

Links in PDF output only work for server-side output.

Therefore, I have updated this to only state it works for PDFs, and if Graphviz is not installed then the function is not available at all.